% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDirichlet.r
\name{calculateConstantCVGenDirichletK}
\alias{calculateConstantCVGenDirichletK}
\title{Calculate Generalized Dirichlet parameters such that the margins have constant Coefficient of Variation}
\usage{
calculateConstantCVGenDirichletK(p, most.likely.k)
}
\arguments{
\item{p}{probabilities}

\item{most.likely.k}{The k of the most likely branch or dirichlet marginal}
}
\value{
the k vector such that all margins have a constant coefficient of variation
}
\description{
Calculate Generalized Dirichlet parameters such that the margins have constant Coefficient of Variation
}
\examples{
calculateConstantCVGenDirichletK
}
