% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.r
\name{calculateDirichletCV}
\alias{calculateDirichletCV}
\title{The coefficient of variation of the marginal betas}
\usage{
calculateDirichletCV(alpha)
}
\arguments{
\item{alpha}{the dirichlet parameters}
}
\value{
the marginal CVs
}
\description{
The coefficient of variation of the marginal betas
}
\examples{
calculateDirichletCV(c(5, 4, 2))
}
