% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDirichlet.r
\name{calculateGenDirichletCV}
\alias{calculateGenDirichletCV}
\title{Calculate the Generalized Dirichlet Coefficient of variation}
\usage{
calculateGenDirichletCV(p, k)
}
\arguments{
\item{p}{vector of probabilities}

\item{k}{vector of parameters}
}
\value{
the coefficient of variation of the marginals
}
\description{
Calculate the Generalized Dirichlet Coefficient of variation
}
\examples{
calculateGenDirichletCV(c(0.5, 0.3, 0.2), c(5,5,3))
}
