% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-package.R
\docType{package}
\name{dirichlet-package}
\alias{dirichlet-package}
\title{Dirichlet Distributions}
\description{
Contains distribution functions for the Dirichlet and Generalized Dirichlet distributions.
}
