% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet}
\alias{ddirichlet}
\alias{dirichlet}
\alias{pdirichlet}
\alias{qdirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

pdirichlet(q, alpha, lower.tail = TRUE, log.p = FALSE)

qdirichlet(p, alpha, lower.tail = TRUE, log.p = FALSE)

rdirichlet(n, alpha)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{alpha}{dirichlet alpha parameter (numeric vector with
positive entries)}

\item{log, log.p}{logical; if TRUE, probabilities p are given as 
log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x] otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is 
taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random 
generation for the dirichlet distribution.
}
\details{
The functions (d/p/q/r)dirichlet simply wrap those of the
standard (d/p/q/r)gamma R implementation, so look at, say, 
\code{\link{dgamma}} for details.
}
\examples{
1 + 1

}
\seealso{
\code{\link{dgamma}}; these functions just wrap the 
  (d/p/q/r)gamma functions.
}

