% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDirichlet.r
\name{gendirichlet}
\alias{gendirichlet}
\alias{dGenDirichlet}
\alias{dGenDirichletStd}
\alias{rGenDirichlet}
\alias{qGenDirichlet}
\title{Generalized Dirichlet functions}
\usage{
dGenDirichlet(x, p, k)

dGenDirichletStd(x, a, b)

rGenDirichlet(n, p, k)

qGenDirichlet(X, p, k)
}
\arguments{
\item{x}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{p, k}{The Generalized Dirichlet normalized parameters.  The Dirichlet parameters may contain zeros if the branch to which a parameter is referring is not allowed.  k * p = alpha}

\item{a, b}{The Generalized Dirichlet standard parameters}

\item{n}{the number of samples}

\item{X}{a matrix of quantiles}
}
\value{
rGenDirichlet generates random deviates and dGenDirichlet generates densities
}
\description{
Random deviates of the Generalized Dirichlet distribution.
}
\examples{
rGenDirichlet(10, c(4,3,2)/9, c(2, 2, 2))
}
