% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.r
\name{qMarginalDirichlet}
\alias{qMarginalDirichlet}
\title{Quantiles of the marginal beta distributions in a dirichlet distribution}
\usage{
qMarginalDirichlet(p, alpha)
}
\arguments{
\item{p}{the probability of the desired quantile}

\item{alpha}{the vector of dirichlet parameters}
}
\value{
the vector of marginal quantiles
}
\description{
Quantiles of the marginal beta distributions in a dirichlet distribution
}
\examples{
qMarginalDirichlet(c(0.1, 0.2, 0.4), c(2,3,4))
}
