\name{dirichlet-package}
\alias{dirichlet-package}
\alias{dirichlet}
\docType{package}
\title{
  Dirichlet Distributions
}
\description{
  Contains distribution functions for the Dirichlet and Generalized Dirichlet distributions
}
\details{
  \tabular{ll}{
    Package: \tab dirichlet\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2013/03/15 16:32:02\cr
    License: \tab GPL >v2\cr
  }
}
\author{
Rob Carnell <bertcarnell@gmail.com>
}
\keyword{distribution}
