% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet}
\alias{ddirichlet}
\alias{dirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

rdirichlet(n, alpha)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{alpha}{dirichlet alpha parameter (numeric vector with
positive entries)}

\item{log, log.p}{logical; if TRUE, probabilities p are given as 
log(p).}

\item{n}{number of observations. If length(n) > 1, the length is 
taken to be the number required.}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x] otherwise, P[X > x].}
}
\description{
Density, distribution function, quantile function and random 
generation for the dirichlet distribution.
}
\details{
The functions (d/p/q/r)dirichlet simply wrap those of the
standard (d/p/q/r)gamma R implementation, so look at, say, 
\code{\link{dgamma}} for details.
}
\examples{

ddirichlet(c(.5,.5), c(.5, .5))
dbeta(.5, .5, .5)

rdirichlet(10, c(.5, .5))

}
\seealso{
\code{\link{dgamma}}; these functions just wrap the 
  (d/p/q/r)gamma functions.
}

